Update creating.lisp
According to spec `(random k)` can return values  0 <= x < k, so `(random 1.0)` occasionally returns 0.0 crashing the `log` function, so I propose to add least-positive-..-float to shift the values. Note that since (random 1.0) < 1.0, adding the least-positive.. cannot cause it to go over 1.0 which would cause a change of sign of the logarithm and complex value of sqrt. The random returning of zero is more pronounced for single floats which was used originally. Also since, according to spec,  `pi` is defined as a long-float, which on most implementations is the same as double-float, the value ends up as double or long float anyways (before the final conversion if requested), so I propose to do the calculations in double floats.